+++ src/ListRenderer/index.ts
@@ -1,5 +1,6 @@
 import { CheckListRenderer } from "./checklistRenderer";
 import { OrderedListRenderer } from "./orderedListRenderer";
 import { UnorderedListRenderer } from "./unorderedListRenderer";
+import { ListRenderer } from './listRenderer';
 
-export default { CheckListRenderer, OrderedListRenderer, UnorderedListRenderer };
+export { CheckListRenderer, OrderedListRenderer, UnorderedListRenderer, ListRenderer };
