+++ src/ListRenderer/listRenderer.ts
@@ -5,6 +5,7 @@ interface NestedListCssClasses {
   wrapper: string;
   wrapperOrdered: string;
   wrapperUnordered: string;
+  wrapperChecklist: string;
   item: string;
   itemBody: string;
   itemContent: string;
@@ -20,7 +21,7 @@ interface NestedListCssClasses {
  * List renderer class
  * Used for storing css classes and
  */
-abstract class ListRenderer {
+export abstract class ListRenderer {
   /**
    * Styles
    *
@@ -32,6 +33,7 @@ abstract class ListRenderer {
       wrapper: 'cdx-nested-list',
       wrapperOrdered: 'cdx-nested-list--ordered',
       wrapperUnordered: 'cdx-nested-list--unordered',
+      wrapperChecklist: 'cdx-nested-list--checklist',
       item: 'cdx-nested-list__item',
       itemBody: 'cdx-nested-list__item-body',
       itemContent: 'cdx-nested-list__item-content',
